backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
pymongo==4.8.0
dnspython==2.6.1
uvicorn==0.30.6
python-dotenv==1.0.1
